bridge en, loop filter, clip exciter
Added a bridge enable (really strike position).
Added a loop filter (always on).
Clipped the beginning of the new exciter so it starts
immediately.
Changed the default velocity if nothing is connected to
the velocity input (at half velocity).
Now at a point to do energy decay reliefs as the sound
is just muddy.